all endpoitns are singleton
