+++ restcomm/restcomm.http/src/main/java/org/restcomm/connect/http/LogoutEndpoint.java
@@ -1,5 +1,6 @@
 package org.restcomm.connect.http;
 
+import com.sun.jersey.spi.resource.Singleton;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -10,6 +11,7 @@ import javax.ws.rs.core.Response;
  * @author <a href="mailto:gvagenas@gmail.com">gvagenas</a>
  */
 @Path("/Logout")
+@Singleton
 public class LogoutEndpoint extends AbstractEndpoint {
 
     @GET
