+++ restcomm/restcomm.http/src/main/java/org/restcomm/connect/http/ProfileEndpoint.java
@@ -26,6 +26,7 @@ import com.github.fge.jsonschema.main.JsonSchema;
 import com.github.fge.jsonschema.main.JsonSchemaFactory;
 import com.sun.jersey.core.header.LinkHeader;
 import com.sun.jersey.core.header.LinkHeader.LinkHeaderBuilder;
+import com.sun.jersey.spi.resource.Singleton;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.MalformedURLException;
@@ -38,12 +39,22 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import javax.annotation.PostConstruct;
+import javax.annotation.security.PermitAll;
+import javax.annotation.security.RolesAllowed;
 import javax.servlet.ServletContext;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.DELETE;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.GenericEntity;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.ResponseBuilder;
 import javax.ws.rs.core.Response.Status;
@@ -53,6 +64,7 @@ import javax.ws.rs.core.UriInfo;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.io.IOUtils;
 import org.apache.log4j.Logger;
+import org.restcomm.connect.commons.annotations.concurrency.ThreadSafe;
 import org.restcomm.connect.commons.dao.Sid;
 import org.restcomm.connect.core.service.api.ProfileService;
 import org.restcomm.connect.dao.AccountsDao;
@@ -67,7 +79,12 @@ import static org.restcomm.connect.dao.entities.Profile.DEFAULT_PROFILE_SID;
 import org.restcomm.connect.dao.entities.ProfileAssociation;
 import org.restcomm.connect.http.exceptionmappers.CustomReasonPhraseType;
 import org.restcomm.connect.http.security.AccountPrincipal;
+import static org.restcomm.connect.http.security.AccountPrincipal.SUPER_ADMIN_ROLE;
 
+@javax.ws.rs.Path("/Profiles")
+@ThreadSafe
+@RolesAllowed(SUPER_ADMIN_ROLE)
+@Singleton
 public class ProfileEndpoint {
 
     protected Logger logger = Logger.getLogger(ProfileEndpoint.class);
@@ -405,5 +422,81 @@ public class ProfileEndpoint {
             logger.debug("getting schema", ex);
             return Response.status(Status.NOT_FOUND).build();
         }
+    }
+
+    private static final String OVERRIDE_HDR = "X-HTTP-Method-Override";
+
+    @GET
+    @Produces(APPLICATION_JSON)
+    public Response getProfilesAsJson(@Context UriInfo info) {
+        return getProfiles(info);
+    }
+
+    @POST
+    @Consumes({PROFILE_CONTENT_TYPE, APPLICATION_JSON})
+    @Produces({PROFILE_CONTENT_TYPE, APPLICATION_JSON})
+    public Response createProfileAsJson(InputStream body, @Context UriInfo info) {
+        return createProfile(body, info);
+    }
+
+    @javax.ws.rs.Path("/{profileSid}")
+    @GET
+    @Produces({PROFILE_CONTENT_TYPE, MediaType.APPLICATION_JSON})
+    @PermitAll
+    public Response getProfileAsJson(@PathParam("profileSid") final String profileSid,
+            @Context UriInfo info, @Context SecurityContext secCtx) {
+        return getProfile(profileSid, info, secCtx);
+    }
+
+    @javax.ws.rs.Path("/{profileSid}")
+    @PUT
+    @Consumes({PROFILE_CONTENT_TYPE, MediaType.APPLICATION_JSON})
+    @Produces({PROFILE_CONTENT_TYPE, MediaType.APPLICATION_JSON})
+    public Response updateProfileAsJson(@PathParam("profileSid") final String profileSid,
+            InputStream body, @Context UriInfo info,
+            @Context HttpHeaders headers) {
+        if (headers.getRequestHeader(OVERRIDE_HDR) != null
+                && headers.getRequestHeader(OVERRIDE_HDR).size() > 0) {
+            String overrideHdr = headers.getRequestHeader(OVERRIDE_HDR).get(0);
+            switch (overrideHdr) {
+                case "LINK":
+                    return linkProfile(profileSid, headers, info);
+                case "UNLINK":
+                    return unlinkProfile(profileSid, headers);
+
+            }
+        }
+        return updateProfile(profileSid, body, info);
+    }
+
+    @javax.ws.rs.Path("/{profileSid}")
+    @DELETE
+    public Response deleteProfileAsJson(@PathParam("profileSid") final String profileSid) {
+        return deleteProfile(profileSid);
+    }
+
+    @javax.ws.rs.Path("/{profileSid}")
+    @LINK
+    @Produces(APPLICATION_JSON)
+    public Response linkProfileAsJson(@PathParam("profileSid") final String profileSid,
+            @Context HttpHeaders headers, @Context UriInfo info
+    ) {
+        return linkProfile(profileSid, headers, info);
+    }
+
+    @javax.ws.rs.Path("/{profileSid}")
+    @UNLINK
+    @Produces(APPLICATION_JSON)
+    public Response unlinkProfileAsJson(@PathParam("profileSid") final String profileSid,
+            @Context HttpHeaders headers) {
+        return unlinkProfile(profileSid, headers);
+    }
+
+    @javax.ws.rs.Path("/schemas/{schemaId}")
+    @GET
+    @Produces({PROFILE_SCHEMA_CONTENT_TYPE, MediaType.APPLICATION_JSON})
+    @PermitAll
+    public Response getProfileSchemaAsJson(@PathParam("schemaId") final String schemaId) {
+        return getSchema(schemaId);
     }
 }
