+++ restcomm/restcomm.http/src/main/java/org/restcomm/connect/http/RecordingsEndpoint.java
@@ -21,6 +21,7 @@ package org.restcomm.connect.http;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.sun.jersey.spi.resource.Singleton;
 import com.thoughtworks.xstream.XStream;
 import java.io.File;
 import java.net.URI;
@@ -76,6 +77,7 @@ import org.restcomm.connect.identity.UserIdentityContext;
  */
 @Path("/Accounts/{accountSid}/Recordings")
 @ThreadSafe
+@Singleton
 public abstract class RecordingsEndpoint extends AbstractEndpoint {
     @Context
     private ServletContext context;
