+++ restcomm/restcomm.http/src/main/java/org/restcomm/connect/http/TranscriptionsEndpoint.java
@@ -21,6 +21,7 @@ package org.restcomm.connect.http;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.sun.jersey.spi.resource.Singleton;
 import com.thoughtworks.xstream.XStream;
 import java.text.ParseException;
 import java.util.ArrayList;
@@ -64,6 +65,7 @@ import org.restcomm.connect.identity.UserIdentityContext;
  * @author quintana.thomas@gmail.com (Thomas Quintana)
  */
 @Path("/Accounts/{accountSid}/Transcriptions")
+@Singleton
 public abstract class TranscriptionsEndpoint extends AbstractEndpoint {
     @Context
     private ServletContext context;
