+++ restcomm/restcomm.http/src/main/java/org/restcomm/connect/http/UsageEndpoint.java
@@ -21,6 +21,7 @@ package org.restcomm.connect.http;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.sun.jersey.spi.resource.Singleton;
 import com.thoughtworks.xstream.XStream;
 import java.util.List;
 import java.util.regex.Pattern;
@@ -65,6 +66,7 @@ import org.restcomm.connect.identity.UserIdentityContext;
  */
 @Path("/Accounts/{accountSid}/Usage/Records")
 @ThreadSafe
+@Singleton
 public abstract class UsageEndpoint extends AbstractEndpoint {
   @Context
   protected ServletContext context;
