+++ restcomm/restcomm.http/src/main/java/org/restcomm/connect/http/UssdPushEndpoint.java
@@ -24,6 +24,7 @@ import static akka.pattern.Patterns.ask;
 import akka.util.Timeout;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.sun.jersey.spi.resource.Singleton;
 import com.thoughtworks.xstream.XStream;
 import java.net.URI;
 import java.util.concurrent.TimeUnit;
@@ -76,6 +77,7 @@ import scala.concurrent.duration.Duration;
  *
  */
 @Path("/Accounts/{accountSid}/UssdPush")
+@Singleton
 public class UssdPushEndpoint extends AbstractEndpoint {
 
     @Context
